requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
drf-spectacular>=0.27
psycopg2-binary>=2.9
python-decouple>=3.8
gunicorn>=22.0
whitenoise>=6.6
cryptography>=42.0
python-json-logger>=2.0
dj-database-url>=2.1
